frontend/requirements.txt
attrs==19.1.0
Click==7.0
csvalidate==1.1.1
Flask==1.0.2
Flask-CSV==1.2.0
Flask-Mako==0.4
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.19.2
Werkzeug==0.15.1
